Renames the WorkspaceWriter and WorkspaceReader to WorkspaceModelWriter and WorkspaceModelReader.
