+++ src/Dyna.Core/Dyna.Core.csproj
@@ -91,8 +91,8 @@
     <Compile Include="Models\VariableDomainExpressionModel.cs" />
     <Compile Include="Models\VariableModel.cs" />
     <Compile Include="Models\WorkspaceModel.cs" />
-    <Compile Include="Models\WorkspaceReader.cs" />
-    <Compile Include="Models\WorkspaceWriter.cs" />
+    <Compile Include="Models\WorkspaceModelReader.cs" />
+    <Compile Include="Models\WorkspaceModelWriter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Solver\OrToolsSolver.cs" />
     <Compile Include="Solver\SolveResult.cs" />
