+++ tests/Dyna.Core.Tests.Integration/Dyna.Core.Tests.Integration.csproj
@@ -82,7 +82,7 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Models\WorkspaceWriterTests.cs" />
+    <Compile Include="Models\WorkspaceModelWriterTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="WorkspaceModelFactory.cs" />
   </ItemGroup>
